requirements.txt
appdirs==1.4.3
cssselect==1.0.1
dj-database-url==0.4.2
Django==1.11
django-ckeditor==5.2.2
gunicorn==19.0.0
Markdown==2.6.8
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
